Networked_Paymenets_open_sourc/requirements.txt
networkx>=3.0
pandas>=2.0
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
plotly>=5.15
streamlit>=1.28
pyyaml>=6.0
openpyxl>=3.1
